fix vue blog

+++ components/blog/ListBlog.vue
@@ -41,6 +41,7 @@
 import axios from "axios";
 import { CATEGORY } from "../../constant/constant";
 import { POSITION } from "../../constant/constant";
+import swal from 'sweetalert2';
 export default {
   name: "list-blog",
   props: {
@@ -65,8 +66,21 @@ export default {
      * @since 18-3-2021
      */
     DeleteBlog(id) {
-      axios.delete("http://localhost:3001/blogs/" + id).then((res) => {
-        this.$emit("getDataBlogAfterDelete", this.dataBlog);
+      swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.delete("http://localhost:3001/blogs/" + id).then((res) => {
+            this.$emit("getDataBlogAfterDelete", this.dataBlog);
+          });
+          swal.fire("Deleted!", "Your file has been deleted.", "success");
+        }
       });
     },
 
@@ -94,25 +108,6 @@ export default {
           return this.POSITION[item];
         })
         .join(",");
-    },
-  },
-  computed: {
-    /** delete Blog by id
-     * @return array CATEGORY
-     *
-     * @since 18-3-2021
-     */
-    getCate() {
-      return this.CATEGORY;
-    },
-
-    /** delete Blog by id
-     * @return array POSITION
-     *
-     * @since 18-3-2021
-     */
-    getPosition() {
-      return this.POSITION;
     },
   },
 };
